Moving shit around (due to being an entry point)
